Integrated Redux reducers and store to work with useEffect hook that makes the API call.Removed unnecessary api.js file and replaced context api with redux store.
