+++ src/components/movie_website/Movie.jsx
@@ -1,6 +1,6 @@
 import React, {useState, useEffect} from 'react'
 import { useParams } from 'react-router'
-import { API_URL } from './api';
+import { API_URL } from './Home';
 import './index.css';
 import { NavLink } from 'react-router-dom';
